+++ R/power_curve_plot.R
@@ -154,6 +154,7 @@ plot.power_curve <- function(x,
 
   what <- match.arg(what, several.ok = TRUE)
 
+  # No need for other arguments. They should be stored in object.
   reject_df <- rejection_rates_add_ci(x$reject_df,
                                       level = ci_level)
 
@@ -255,8 +256,9 @@ plot.power4test_by_n <- function(
                              ci_level = .95,
                              ...) {
   what <- match.arg(what, several.ok = TRUE)
+  # No need for other arguments. They should be stored in object.
   reject_df <- rejection_rates(x,
-                               ci_level = ci_level,
+                               level = ci_level,
                                all_columns = TRUE)
   reject_df$x <- reject_df$n
   x0 <- list(predictor = "n",
@@ -275,37 +277,6 @@ plot.power4test_by_n <- function(
   invisible(x)
 }
 
-# #' @rdname plot.power_curve
-# #' @export
-# plot.power4test_by_n <- function(
-#                              x,
-#                              main = "Estimated Power vs. Sample Size",
-#                              xlab = "Sample Size",
-#                              ylab = "Estimated Power",
-#                              pars_ci = list(),
-#                              type = "l",
-#                              ylim = c(0, 1),
-#                              ci_level = .95,
-#                              ...) {
-#   reject_df <- rejection_rates(x,
-#                                ci_level = ci_level,
-#                                all_columns = TRUE)
-#   reject_df$x <- reject_df$n
-#   x0 <- list(predictor = "n",
-#              reject_df = reject_df)
-
-#   plot.power_curve(x = x0,
-#                    main = main,
-#                    xlab = xlab,
-#                    ylab = ylab,
-#                    pars_ci = pars_ci,
-#                    type = type,
-#                    ylim = ylim,
-#                    ci_level = ci_level,
-#                    ...)
-#   invisible(x)
-# }
-
 #' @rdname plot.power_curve
 #' @export
 plot.power4test_by_es <- function(
@@ -322,8 +293,9 @@ plot.power4test_by_es <- function(
                              ci_level = .95,
                              ...) {
   what <- match.arg(what, several.ok = TRUE)
+  # No need for other arguments. They should be stored in object.
   reject_df <- rejection_rates(x,
-                               ci_level = ci_level,
+                               level = ci_level,
                                all_columns = TRUE)
   reject_df$x <- reject_df$es
   x0 <- list(predictor = "es",