+++ R/x_from_power_for_es_helpers.R
@@ -10,6 +10,7 @@ set_es_range_by_x <- function(
                           goal = NULL,
                           tol = NULL,
                           ci_level = .95) {
+  # Other arguments will be retrieved from object_by_org
   reject0by <- rejection_rates(object_by_org,
                                level = ci_level,
                                add_se = TRUE,
@@ -72,6 +73,7 @@ set_es_range <- function(object,
                 pop_es_name = pop_es_name)
   es0_sign <- sign(es0)
   es0_abs <- abs(es0)
+  # No need for other arguments because only `reject` is used
   reject0 <- rejection_rates(object)
   power0 <- reject0$reject[1]
   if (es0 > es_max) {