+++ R/x_from_power_for_x_plot.R
@@ -435,6 +435,7 @@ plot_power_x <- function(object,
                          ylim = c(0, 1),
                          ci_level = .95,
                          ...) {
+  # No need for other arguments. They should be stored in object.
   reject0 <- rejection_rates_add_ci(object,
                                     level = ci_level)
   x <- switch(predictor,
@@ -458,6 +459,7 @@ plot_power_x_ci <- function(object,
                             col = "grey50",
                             ci_level = .95,
                             ...) {
+  # No need for other arguments. They should be stored in object.
   reject0 <- rejection_rates_add_ci(object,
                                     level = ci_level)
   x <- switch(predictor,
@@ -495,6 +497,7 @@ plot_power_curve_x <- function(object,
                                ci_level = .95,
                                ...) {
   # power_x_fit is a power_curve object
+  # No need for other arguments. They should be stored in object.
   reject0 <- rejection_rates_add_ci(object,
                                     level = ci_level)
   x <- switch(predictor,
@@ -591,6 +594,7 @@ plot.n_region_from_power <- function(x,
   # It is intended *not* to use plot.power_curve().
   # It is possible that the fit failed.
 
+  # No need for other arguments. They should be stored in object.
   a1 <- rejection_rates(
               x$below$power4test_trials,
               all_columns = TRUE
@@ -643,14 +647,6 @@ plot.n_region_from_power <- function(x,
 
   do.call(plot_power_x,
           args1)
-
-  # do.call(plot_power_x,
-  #         list(object = a,
-  #              predictor = predictor,
-  #              main = main,
-  #              xlab = xlab,
-  #              ylab = ylab,
-  #              ...))
 
   # === Add CIs?