+++ tests/testthat/test-x_from_power_algorithm_bisection_short.R
@@ -1,3 +1,7 @@
+skip("Not used")
+
+# Do not use these tests. They are "too low level."
+
 skip_on_cran()
 
 library(testthat)
@@ -52,12 +56,12 @@ a_out <- power_algorithm_bisection(object = out,
                                    by_x_1 = by_x_1,
                                    x_interval = c(600, 700),
                                    goal = "close_enough",
-                                   tol = .04,
+                                   tol = .07,
                                    final_nrep = 20,
                                    progress = !is_testing(),
                                    simulation_progress = !is_testing())
 rejection_rates(a_out$by_x_1)
-expect_true(abs(a_out$power_out - .80) < .04)
+expect_true(abs(a_out$power_out - .80) < .07)
 
 set.seed(1234)
 a_out <- power_algorithm_bisection(object = out,